헤더 및 게시글들 폴더 정리
