+++ frontend/src/App.js
@@ -1,5 +1,5 @@
 import React from "react";
-import DefaultPage from "./component/DefaultPage";
+import DefaultPage from "./component/main/DefaultPage";
 import {BrowserRouter as Router, Route, Routes,} from "react-router-dom";
 
 function App() {
